+++ src/main/java/com/agile/vision/summer/practice/services/PlaceServiceImpl.java
@@ -6,6 +6,7 @@ import com.agile.vision.summer.practice.entities.Place;
 import com.agile.vision.summer.practice.repositories.MonitorRepository;
 import com.agile.vision.summer.practice.repositories.PcRepository;
 import com.agile.vision.summer.practice.repositories.WorkingPlaceRepository;
+import com.agile.vision.summer.practice.services.exception.NonExistingIdException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,7 +29,7 @@ public class PlaceServiceImpl implements PlaceService {
         if (optional.isPresent()) {
             return optional.get();
         }
-        throw new IllegalArgumentException();
+        throw new NonExistingIdException();
     }
 
     @Override
@@ -38,11 +39,8 @@ public class PlaceServiceImpl implements PlaceService {
 
     @Override
     public boolean deleteById(int id) {
-        try {
-            workingPlaceRepository.deleteById(id);
-        } catch (Exception e) {
-            return false;
-        }
+        if (workingPlaceRepository.existsById(id)) throw new NonExistingIdException();
+        workingPlaceRepository.deleteById(id);
         return true;
     }
 
@@ -53,11 +51,7 @@ public class PlaceServiceImpl implements PlaceService {
 
     @Override
     public boolean deleteAllByPcNullOrMonitorsIsNull() {
-        try {
-            workingPlaceRepository.deleteAllByPcNullOrMonitorsIsNull();
-        } catch(Exception e){
-            return false;
-        }
+        workingPlaceRepository.deleteAllByPcNullOrMonitorsIsNull();
         return true;
     }
 }
